backend/requirements.txt
fastapi
uvicorn

pymupdf
pdfplumber
pytesseract
pillow

sentence-transformers
faiss-cpu
numpy

openai
python-dotenv
python-multipart

unstructured[pdf]
unstructured-inference

opencv-python-headless
pandas